Change breadcrumbs to use pushstate via Link
If you tried to navigate from the ChordEditor back to the songs list
by click on a breadcrumb it would cause an actual redirect instead of
using pushstate. This would cause the application to need to load again and
would additionally cause a 404 if the sub-path redirection isn't present on
your webserver.